Make the message builders inherit more things
