+++ DSharpPlus/Entities/BaseDiscordMessageBuilder.cs
@@ -21,8 +21,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 
 namespace DSharpPlus.Entities
 {
@@ -33,12 +35,46 @@ namespace DSharpPlus.Entities
     public abstract class BaseDiscordMessageBuilder<T> : IDiscordMessageBuilder where T : BaseDiscordMessageBuilder<T>
         // This has got to be the most big brain thing I have ever done with interfaces lmfao
     {
-        public abstract string Content { get; set; }
-        public abstract bool IsTTS { get; set; }
-        public abstract IReadOnlyList<DiscordEmbed> Embeds { get; }
-        public abstract IReadOnlyList<DiscordMessageFile> Files { get; }
-        public abstract IReadOnlyList<DiscordActionRowComponent> Components { get; }
-        public abstract IReadOnlyList<IMention> Mentions { get; }
+        /// <summary>
+        /// Message to send on this webhook request.
+        /// </summary>
+        public string Content
+        {
+            get => this._content;
+            set
+            {
+                if (value != null && value.Length > 2000)
+                    throw new ArgumentException("Content length cannot exceed 2000 characters.", nameof(value));
+                this._content = value;
+            }
+        }
+        protected string _content;
+
+        public bool IsTTS { get; set; }
+
+        /// <summary>
+        /// Embeds to send on this webhook request.
+        /// </summary>
+        public IReadOnlyList<DiscordEmbed> Embeds => this._embeds;
+        protected List<DiscordEmbed> _embeds = new();
+
+        /// <summary>
+        /// Files to send on this webhook request.
+        /// </summary>
+        public IReadOnlyList<DiscordMessageFile> Files => this._files;
+        protected List<DiscordMessageFile> _files = new();
+
+        /// <summary>
+        /// Mentions to send on this webhook request.
+        /// </summary>
+        public IReadOnlyList<IMention> Mentions => this._mentions;
+        protected List<IMention> _mentions = new();
+
+        /// <summary>
+        /// Components to send on this followup message.
+        /// </summary>
+        public IReadOnlyList<DiscordActionRowComponent> Components => this._components;
+        protected List<DiscordActionRowComponent> _components = new();
 
         /// <summary>
         /// Thou shalt NOT PASS! ⚡
@@ -46,102 +82,232 @@ namespace DSharpPlus.Entities
         internal BaseDiscordMessageBuilder(){}
 
         /// <summary>
-        /// Adds content to this message
+        /// Constructs a new <see cref="BaseDiscordMessageBuilder{T}"/> based on an existing <see cref="IDiscordMessageBuilder"/>.
+        /// Existing file streams will have their position reset to 0.
         /// </summary>
-        /// <param name="content">Message content to use</param>
+        /// <param name="builder">The builder to copy.</param>
+        public BaseDiscordMessageBuilder(IDiscordMessageBuilder builder)
+        {
+            this._content = builder.Content;
+            this._mentions.AddRange(builder.Mentions.ToList());
+            this._embeds.AddRange(builder.Embeds);
+            this._components.AddRange(builder.Components);
+            this._files.AddRange(builder.Files);
+            this.IsTTS = builder.IsTTS;
+        }
+
+        /// <summary>
+        /// Sets the Content of the Message.
+        /// </summary>
+        /// <param name="content">The content to be set.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T WithContent(string content)
+        {
+            this.Content = content;
+            return this as T;
+        }
+
+        /// <summary>
+        /// Adds a row of components to a message, up to 5 components per row, and up to 5 rows per message.
+        /// </summary>
+        /// <param name="components">The components to add to the message.</param>
+        /// <returns>The current builder to be chained.</returns>
+        /// <exception cref="ArgumentOutOfRangeException">No components were passed.</exception>
+        public T AddComponents(params DiscordComponent[] components)
+            => this.AddComponents((IEnumerable<DiscordComponent>)components);
+
+
+        /// <summary>
+        /// Appends several rows of components to the message
+        /// </summary>
+        /// <param name="components">The rows of components to add, holding up to five each.</param>
         /// <returns></returns>
-        public abstract T WithContent(string content);
+        public T AddComponents(IEnumerable<DiscordActionRowComponent> components)
+        {
+            var ara = components.ToArray();
+
+            if (ara.Length + this._components.Count > 5)
+                throw new ArgumentException("ActionRow count exceeds maximum of five.");
+
+            foreach (var ar in ara)
+                this._components.Add(ar);
+
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds components to this message. Each call should append to a new row.
+        /// Adds a row of components to a message, up to 5 components per row, and up to 5 rows per message.
         /// </summary>
-        /// <param name="components">Components to add.</param>
-        /// <returns></returns>
-        public abstract T AddComponents(params DiscordComponent[] components);
+        /// <param name="components">The components to add to the message.</param>
+        /// <returns>The current builder to be chained.</returns>
+        /// <exception cref="ArgumentOutOfRangeException">No components were passed.</exception>
+        public T AddComponents(IEnumerable<DiscordComponent> components)
+        {
+            var cmpArr = components.ToArray();
+            var count = cmpArr.Length;
+
+            if (!cmpArr.Any())
+                throw new ArgumentOutOfRangeException(nameof(components), "You must provide at least one component");
+
+            if (count > 5)
+                throw new ArgumentException("Cannot add more than 5 components per action row!");
+
+            var comp = new DiscordActionRowComponent(cmpArr);
+            this._components.Add(comp);
+
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds components to this message. Each call should append to a new row.
+        /// Sets if the message should be TTS.
         /// </summary>
-        /// <param name="components">Components to add.</param>
-        /// <returns></returns>
-        public abstract T AddComponents(IEnumerable<DiscordComponent> components);
+        /// <param name="isTTS">If TTS should be set.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T WithTTS(bool isTTS)
+        {
+            this.IsTTS = isTTS;
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds an action row component to this message.
+        /// Appends an embed to the current builder.
         /// </summary>
-        /// <param name="components">Action row to add to this message. Should contain child components.</param>
-        /// <returns></returns>
-        public abstract T AddComponents(IEnumerable<DiscordActionRowComponent> components);
+        /// <param name="embed">The embed that should be appended.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T AddEmbed(DiscordEmbed embed)
+        {
+            if (embed == null)
+                return this as T; //Providing null embeds will produce a 400 response from Discord.//
+            this._embeds.Add(embed);
+            return this as T;
+        }
 
         /// <summary>
-        /// Sets whether this message should play as a text-to-speech message.
+        /// Appends several embeds to the current builder.
         /// </summary>
-        /// <param name="isTTS"></param>
-        /// <returns></returns>
-        public abstract T WithTTS(bool isTTS);
+        /// <param name="embeds">The embeds that should be appended.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T AddEmbeds(IEnumerable<DiscordEmbed> embeds)
+        {
+            this._embeds.AddRange(embeds);
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds an embed to this message.
+        /// Sets if the message has files to be sent.
         /// </summary>
-        /// <param name="embed">Embed to add.</param>
-        /// <returns></returns>
-        public abstract T AddEmbed(DiscordEmbed embed);
+        /// <param name="fileName">The fileName that the file should be sent as.</param>
+        /// <param name="stream">The Stream to the file.</param>
+        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T AddFile(string fileName, Stream stream, bool resetStreamPosition = false)
+        {
+            if (this.Files.Count >= 10)
+                throw new ArgumentException("Cannot send more than 10 files with a single message.");
+
+            if (this._files.Any(x => x.FileName == fileName))
+                throw new ArgumentException("A File with that filename already exists");
+
+            if (resetStreamPosition)
+                this._files.Add(new DiscordMessageFile(fileName, stream, stream.Position));
+            else
+                this._files.Add(new DiscordMessageFile(fileName, stream, null));
+
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds multiple embeds to this message.
+        /// Sets if the message has files to be sent.
         /// </summary>
-        /// <param name="embeds">Collection of embeds to add.</param>
-        /// <returns></returns>
-        public abstract T AddEmbeds(IEnumerable<DiscordEmbed> embeds);
+        /// <param name="stream">The Stream to the file.</param>
+        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T AddFile(FileStream stream, bool resetStreamPosition = false)
+        {
+            if (this.Files.Count >= 10)
+                throw new ArgumentException("Cannot send more than 10 files with a single message.");
+
+            if (this._files.Any(x => x.FileName == stream.Name))
+                throw new ArgumentException("A File with that filename already exists");
+
+            if (resetStreamPosition)
+                this._files.Add(new DiscordMessageFile(stream.Name, stream, stream.Position));
+            else
+                this._files.Add(new DiscordMessageFile(stream.Name, stream, null));
+
+            return this as T;
+        }
 
         /// <summary>
-        /// Attaches a file to this message.
+        /// Sets if the message has files to be sent.
         /// </summary>
-        /// <param name="fileName">Name of the file to attach.</param>
-        /// <param name="stream">Stream containing said file's contents.</param>
-        /// <param name="resetStream">Whether to reset the stream to position 0 after sending.</param>
-        /// <returns></returns>
-        public abstract T AddFile(string fileName, Stream stream, bool resetStream = false);
+        /// <param name="files">The Files that should be sent.</param>
+        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
+        /// <returns>The current builder to be chained.</returns>
+        public T AddFiles(IDictionary<string, Stream> files, bool resetStreamPosition = false)
+        {
+            if (this.Files.Count + files.Count > 10)
+                throw new ArgumentException("Cannot send more than 10 files with a single message.");
+
+            foreach (var file in files)
+            {
+                if (this._files.Any(x => x.FileName == file.Key))
+                    throw new ArgumentException("A File with that filename already exists");
+
+                if (resetStreamPosition)
+                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, file.Value.Position));
+                else
+                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, null));
+            }
+
+            return this as T;
+        }
+
+        public T AddFiles(IEnumerable<DiscordMessageFile> files)
+        {
+            this._files.AddRange(files);
+            return this as T;
+        }
+
 
         /// <summary>
-        /// Attaches a file to this message.
+        /// Adds the mention to the mentions to parse, etc. with the interaction response.
         /// </summary>
-        /// <param name="stream">FileStream pointiong to the file to attach.</param>
-        /// <param name="resetStream">Whether to reset the stream position to 0 after sending.</param>
-        /// <returns></returns>
-        public abstract T AddFile(FileStream stream, bool resetStream = false);
+        /// <param name="mention">Mention to add.</param>
+        public T AddMention(IMention mention)
+        {
+            this._mentions.Add(mention);
+            return this as T;
+        }
 
         /// <summary>
-        /// Attaches multiple files to this message.
+        /// Adds the mentions to the mentions to parse, etc. with the interaction response.
         /// </summary>
-        /// <param name="files">Dictionary of files to add, where <see cref="string"/> is a file name and <see cref="Stream"/> is a stream containing the file's contents.</param>
-        /// <param name="resetStreams">Whether to reset all stream positions to 0 after sending.</param>
-        /// <returns></returns>
-        public abstract T AddFiles(IDictionary<string, Stream> files, bool resetStreams = false);
+        /// <param name="mentions">Mentions to add.</param>
+        public T AddMentions(IEnumerable<IMention> mentions)
+        {
+            this._mentions.AddRange(mentions);
+            return this as T;
+        }
 
         /// <summary>
-        /// Adds an allowed mention to this message.
+        /// Clears all message components on this builder.
         /// </summary>
-        /// <param name="mention">Mention to allow in this message.</param>
-        /// <returns></returns>
-        public abstract T AddMention(IMention mention);
+        public virtual void ClearComponents()
+            => this._components.Clear();
 
         /// <summary>
-        /// Adds multiple allowed mentions to this message.
+        /// Allows for clearing the Message Builder so that it can be used again to send a new message.
         /// </summary>
-        /// <param name="mentions">Collection of mentions to allow in this message.</param>
-        /// <returns></returns>
-        public abstract T AddMentions(IEnumerable<IMention> mentions);
-
-        /// <summary>
-        /// Clears all components attached to this builder.
-        /// </summary>
-        public abstract void ClearComponents();
-
-        /// <summary>
-        /// Clears this builder.
-        /// </summary>
-        public abstract void Clear();
+        public virtual void Clear()
+        {
+            this.Content = "";
+            this._embeds.Clear();
+            this.IsTTS = false;
+            this._mentions.Clear();
+            this._files.Clear();
+            this._components.Clear();
+        }
 
 
         IDiscordMessageBuilder IDiscordMessageBuilder.WithContent(string content) => this.WithContent(content);
@@ -154,6 +320,7 @@ namespace DSharpPlus.Entities
         IDiscordMessageBuilder IDiscordMessageBuilder.AddFile(string fileName, Stream stream, bool resetStream) => this.AddFile(fileName, stream, resetStream);
         IDiscordMessageBuilder IDiscordMessageBuilder.AddFile(FileStream stream, bool resetStream) => this.AddFile(stream, resetStream);
         IDiscordMessageBuilder IDiscordMessageBuilder.AddFiles(IDictionary<string, Stream> files, bool resetStreams) => this.AddFiles(files, resetStreams);
+        IDiscordMessageBuilder IDiscordMessageBuilder.AddFiles(IEnumerable<DiscordMessageFile> files) => this.AddFiles(files);
         IDiscordMessageBuilder IDiscordMessageBuilder.AddMention(IMention mention) => this.AddMention(mention);
         IDiscordMessageBuilder IDiscordMessageBuilder.AddMentions(IEnumerable<IMention> mentions) => this.AddMentions(mentions);
     }
@@ -263,6 +430,13 @@ namespace DSharpPlus.Entities
         /// <param name="resetStreams">Whether to reset all stream positions to 0 after sending.</param>
         /// <returns></returns>
         IDiscordMessageBuilder AddFiles(IDictionary<string, Stream> files, bool resetStreams = false);
+
+        /// <summary>
+        /// Attaches previously used files to this file stream.
+        /// </summary>
+        /// <param name="files">Previously attached files to reattach</param>
+        /// <returns></returns>
+        IDiscordMessageBuilder AddFiles(IEnumerable<DiscordMessageFile> files);
 
         /// <summary>
         /// Adds an allowed mention to this message.
