+++ DSharpPlus/Entities/Interaction/DiscordFollowupMessageBuilder.cs
@@ -22,8 +22,6 @@
 // SOFTWARE.
 
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 
 namespace DSharpPlus.Entities
@@ -34,239 +32,23 @@ namespace DSharpPlus.Entities
     public sealed class DiscordFollowupMessageBuilder : BaseDiscordMessageBuilder<DiscordFollowupMessageBuilder>
     {
         /// <summary>
-        /// Whether this followup message is text-to-speech.
-        /// </summary>
-        public override bool IsTTS { get; set; }
-
-        /// <summary>
         /// Whether this followup message should be ephemeral.
         /// </summary>
         public bool IsEphemeral { get; set; }
 
-        internal int? Flags
+        internal int? _flags
             => this.IsEphemeral ? 64 : null;
 
         /// <summary>
-        /// Message to send on followup message.
+        /// Constructs a new followup message builder
         /// </summary>
-        public override string Content
-        {
-            get => this._content;
-            set
-            {
-                if (value != null && value.Length > 2000)
-                    throw new ArgumentException("Content length cannot exceed 2000 characters.", nameof(value));
-                this._content = value;
-            }
-        }
-        private string _content;
+        public DiscordFollowupMessageBuilder() { }
 
         /// <summary>
-        /// Embeds to send on followup message.
+        /// Constructs a new followup message builder based on a previous IDiscordMessageBuilder
         /// </summary>
-        public override IReadOnlyList<DiscordEmbed> Embeds => this._embeds;
-        private readonly List<DiscordEmbed> _embeds = new();
-
-        /// <summary>
-        /// Files to send on this followup message.
-        /// </summary>
-        public override IReadOnlyList<DiscordMessageFile> Files => this._files;
-        private readonly List<DiscordMessageFile> _files = new();
-
-        /// <summary>
-        /// Components to send on this followup message.
-        /// </summary>
-        public override IReadOnlyList<DiscordActionRowComponent> Components => this._components;
-        private readonly List<DiscordActionRowComponent> _components = new();
-
-        /// <summary>
-        /// Mentions to send on this followup message.
-        /// </summary>
-        public override IReadOnlyList<IMention> Mentions => this._mentions;
-        private readonly List<IMention> _mentions = new();
-
-
-        /// <summary>
-        /// Appends a collection of components to the message.
-        /// </summary>
-        /// <param name="components">The collection of components to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        /// <exception cref="ArgumentException"><paramref name="components"/> contained more than 5 components.</exception>
-        public override DiscordFollowupMessageBuilder AddComponents(params DiscordComponent[] components)
-            => this.AddComponents((IEnumerable<DiscordComponent>)components);
-
-        /// <summary>
-        /// Appends several rows of components to the message
-        /// </summary>
-        /// <param name="components">The rows of components to add, holding up to five each.</param>
-        /// <returns></returns>
-        public override DiscordFollowupMessageBuilder AddComponents(IEnumerable<DiscordActionRowComponent> components)
-        {
-            var ara = components.ToArray();
-
-            if (ara.Length + this._components.Count > 5)
-                throw new ArgumentException("ActionRow count exceeds maximum of five.");
-
-            foreach (var ar in ara)
-                this._components.Add(ar);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Appends a collection of components to the message.
-        /// </summary>
-        /// <param name="components">The collection of components to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        /// <exception cref="ArgumentException"><paramref name="components"/> contained more than 5 components.</exception>
-        public override DiscordFollowupMessageBuilder AddComponents(IEnumerable<DiscordComponent> components)
-        {
-            var compArr = components.ToArray();
-            var count = compArr.Length;
-
-            if (count > 5)
-                throw new ArgumentException("Cannot add more than 5 components per action row!");
-
-            var arc = new DiscordActionRowComponent(compArr);
-            this._components.Add(arc);
-            return this;
-        }
-        /// <summary>
-        /// Indicates if the followup message must use text-to-speech.
-        /// </summary>
-        /// <param name="tts">Text-to-speech</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder WithTTS(bool tts)
-        {
-            this.IsTTS = tts;
-            return this;
-        }
-
-        /// <summary>
-        /// Sets the message to send with the followup message..
-        /// </summary>
-        /// <param name="content">Message to send.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder WithContent(string content)
-        {
-            this.Content = content;
-            return this;
-        }
-
-        /// <summary>
-        /// Adds an embed to the followup message.
-        /// </summary>
-        /// <param name="embed">Embed to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddEmbed(DiscordEmbed embed)
-        {
-            this._embeds.Add(embed);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the given embeds to the followup message.
-        /// </summary>
-        /// <param name="embeds">Embeds to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddEmbeds(IEnumerable<DiscordEmbed> embeds)
-        {
-            this._embeds.AddRange(embeds);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds a file to the followup message.
-        /// </summary>
-        /// <param name="filename">Name of the file.</param>
-        /// <param name="data">File data.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddFile(string filename, Stream data, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count >= 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            if (this._files.Any(x => x.FileName == filename))
-                throw new ArgumentException("A File with that filename already exists");
-
-            if (resetStreamPosition)
-                this._files.Add(new DiscordMessageFile(filename, data, data.Position));
-            else
-                this._files.Add(new DiscordMessageFile(filename, data, null));
-
-            return this;
-        }
-
-        /// <summary>
-        /// Sets if the message has files to be sent.
-        /// </summary>
-        /// <param name="stream">The Stream to the file.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddFile(FileStream stream, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count >= 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            if (this._files.Any(x => x.FileName == stream.Name))
-                throw new ArgumentException("A File with that filename already exists");
-
-            if (resetStreamPosition)
-                this._files.Add(new DiscordMessageFile(stream.Name, stream, stream.Position));
-            else
-                this._files.Add(new DiscordMessageFile(stream.Name, stream, null));
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the given files to the followup message.
-        /// </summary>
-        /// <param name="files">Dictionary of file name and file data.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddFiles(IDictionary<string, Stream> files, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count + files.Count > 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            foreach (var file in files)
-            {
-                if (this._files.Any(x => x.FileName == file.Key))
-                    throw new ArgumentException("A File with that filename already exists");
-
-                if (resetStreamPosition)
-                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, file.Value.Position));
-                else
-                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, null));
-            }
-
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the mention to the mentions to parse, etc. with the followup message.
-        /// </summary>
-        /// <param name="mention">Mention to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddMention(IMention mention)
-        {
-            this._mentions.Add(mention);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the mentions to the mentions to parse, etc. with the followup message.
-        /// </summary>
-        /// <param name="mentions">Mentions to add.</param>
-        /// <returns>The builder to chain calls with.</returns>
-        public override DiscordFollowupMessageBuilder AddMentions(IEnumerable<IMention> mentions)
-        {
-            this._mentions.AddRange(mentions);
-            return this;
-        }
+        /// <param name="builder"></param>
+        public DiscordFollowupMessageBuilder(IDiscordMessageBuilder builder) : base(builder) { }
 
         /// <summary>
         /// Sets the followup message to be ephemeral.
@@ -280,23 +62,13 @@ namespace DSharpPlus.Entities
         }
 
         /// <summary>
-        /// Clears all message components on this builder.
-        /// </summary>
-        public override void ClearComponents()
-            => this._components.Clear();
-
-        /// <summary>
         /// Allows for clearing the Followup Message builder so that it can be used again to send a new message.
         /// </summary>
         public override void Clear()
         {
-            this.Content = "";
-            this._embeds.Clear();
-            this.IsTTS = false;
-            this._mentions.Clear();
-            this._files.Clear();
             this.IsEphemeral = false;
-            this._components.Clear();
+
+            base.Clear();
         }
 
         internal void Validate()
