+++ DSharpPlus/Entities/Webhook/DiscordWebhookBuilder.cs
@@ -22,8 +22,6 @@
 // SOFTWARE.
 
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -45,110 +43,20 @@ namespace DSharpPlus.Entities
         public Optional<string> AvatarUrl { get; set; }
 
         /// <summary>
-        /// Whether this webhook request is text-to-speech.
-        /// </summary>
-        public override bool IsTTS { get; set; }
-
-        /// <summary>
-        /// Message to send on this webhook request.
-        /// </summary>
-        public override string Content
-        {
-            get => this._content;
-            set
-            {
-                if (value != null && value.Length > 2000)
-                    throw new ArgumentException("Content length cannot exceed 2000 characters.", nameof(value));
-                this._content = value;
-            }
-        }
-
-        private string _content;
-
-        /// <summary>
         /// Id of the thread to send the webhook request to.
         /// </summary>
         public ulong? ThreadId { get; set; }
-
-        /// <summary>
-        /// Embeds to send on this webhook request.
-        /// </summary>
-        public override IReadOnlyList<DiscordEmbed> Embeds => this._embeds;
-        private readonly List<DiscordEmbed> _embeds = new();
-
-        /// <summary>
-        /// Files to send on this webhook request.
-        /// </summary>
-        public override IReadOnlyList<DiscordMessageFile> Files => this._files;
-        private readonly List<DiscordMessageFile> _files = new();
-
-        /// <summary>
-        /// Mentions to send on this webhook request.
-        /// </summary>
-        public override IReadOnlyList<IMention> Mentions => this._mentions;
-        private readonly List<IMention> _mentions = new();
-
-
-        public override IReadOnlyList<DiscordActionRowComponent> Components => this._components;
-        private readonly List<DiscordActionRowComponent> _components = new();
-
 
         /// <summary>
         /// Constructs a new empty webhook request builder.
         /// </summary>
         public DiscordWebhookBuilder() { } // I still see no point in initializing collections with empty collections. //
 
-
         /// <summary>
-        /// Adds a row of components to the builder, up to 5 components per row, and up to 5 rows per message.
+        /// Constructs a new webhook request builder based on a previous message builder
         /// </summary>
-        /// <param name="components">The components to add to the builder.</param>
-        /// <returns>The current builder to be chained.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">No components were passed.</exception>
-        public override DiscordWebhookBuilder AddComponents(params DiscordComponent[] components)
-            => this.AddComponents((IEnumerable<DiscordComponent>)components);
-
-
-        /// <summary>
-        /// Appends several rows of components to the builder
-        /// </summary>
-        /// <param name="components">The rows of components to add, holding up to five each.</param>
-        /// <returns></returns>
-        public override DiscordWebhookBuilder AddComponents(IEnumerable<DiscordActionRowComponent> components)
-        {
-            var ara = components.ToArray();
-
-            if (ara.Length + this._components.Count > 5)
-                throw new ArgumentException("ActionRow count exceeds maximum of five.");
-
-            foreach (var ar in ara)
-                this._components.Add(ar);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds a row of components to the builder, up to 5 components per row, and up to 5 rows per message.
-        /// </summary>
-        /// <param name="components">The components to add to the builder.</param>
-        /// <returns>The current builder to be chained.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">No components were passed.</exception>
-        public override DiscordWebhookBuilder AddComponents(IEnumerable<DiscordComponent> components)
-        {
-            var cmpArr = components.ToArray();
-            var count = cmpArr.Length;
-
-            if (!cmpArr.Any())
-                throw new ArgumentOutOfRangeException(nameof(components), "You must provide at least one component");
-
-            if (count > 5)
-                throw new ArgumentException("Cannot add more than 5 components per action row!");
-
-            var comp = new DiscordActionRowComponent(cmpArr);
-            this._components.Add(comp);
-
-            return this;
-        }
+        /// <param name="builder"></param>
+        public DiscordWebhookBuilder(IDiscordMessageBuilder builder) : base(builder) { }
 
         /// <summary>
         /// Sets the username for this webhook builder.
@@ -167,137 +75,6 @@ namespace DSharpPlus.Entities
         public DiscordWebhookBuilder WithAvatarUrl(string avatarUrl)
         {
             this.AvatarUrl = avatarUrl;
-            return this;
-        }
-
-        /// <summary>
-        /// Indicates if the webhook must use text-to-speech.
-        /// </summary>
-        /// <param name="tts">Text-to-speech</param>
-        public override DiscordWebhookBuilder WithTTS(bool tts)
-        {
-            this.IsTTS = tts;
-            return this;
-        }
-
-        /// <summary>
-        /// Sets the message to send at the execution of the webhook.
-        /// </summary>
-        /// <param name="content">Message to send.</param>
-        public override DiscordWebhookBuilder WithContent(string content)
-        {
-            this.Content = content;
-            return this;
-        }
-
-        /// <summary>
-        /// Adds an embed to send at the execution of the webhook.
-        /// </summary>
-        /// <param name="embed">Embed to add.</param>
-        public override DiscordWebhookBuilder AddEmbed(DiscordEmbed embed)
-        {
-            if (embed != null)
-                this._embeds.Add(embed);
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the given embeds to send at the execution of the webhook.
-        /// </summary>
-        /// <param name="embeds">Embeds to add.</param>
-        public override DiscordWebhookBuilder AddEmbeds(IEnumerable<DiscordEmbed> embeds)
-        {
-            this._embeds.AddRange(embeds);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds a file to send at the execution of the webhook.
-        /// </summary>
-        /// <param name="filename">Name of the file.</param>
-        /// <param name="data">File data.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        public override DiscordWebhookBuilder AddFile(string filename, Stream data, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count() >= 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            if (this._files.Any(x => x.FileName == filename))
-                throw new ArgumentException("A File with that filename already exists");
-
-            if (resetStreamPosition)
-                this._files.Add(new DiscordMessageFile(filename, data, data.Position));
-            else
-                this._files.Add(new DiscordMessageFile(filename, data, null));
-
-            return this;
-        }
-
-        /// <summary>
-        /// Sets if the message has files to be sent.
-        /// </summary>
-        /// <param name="stream">The Stream to the file.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        /// <returns></returns>
-        public override DiscordWebhookBuilder AddFile(FileStream stream, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count() >= 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            if (this._files.Any(x => x.FileName == stream.Name))
-                throw new ArgumentException("A File with that filename already exists");
-
-            if (resetStreamPosition)
-                this._files.Add(new DiscordMessageFile(stream.Name, stream, stream.Position));
-            else
-                this._files.Add(new DiscordMessageFile(stream.Name, stream, null));
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the given files to send at the execution of the webhook.
-        /// </summary>
-        /// <param name="files">Dictionary of file name and file data.</param>
-        /// <param name="resetStreamPosition">Tells the API Client to reset the stream position to what it was after the file is sent.</param>
-        public override DiscordWebhookBuilder AddFiles(IDictionary<string, Stream> files, bool resetStreamPosition = false)
-        {
-            if (this.Files.Count() + files.Count() > 10)
-                throw new ArgumentException("Cannot send more than 10 files with a single message.");
-
-            foreach (var file in files)
-            {
-                if (this._files.Any(x => x.FileName == file.Key))
-                    throw new ArgumentException("A File with that filename already exists");
-
-                if (resetStreamPosition)
-                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, file.Value.Position));
-                else
-                    this._files.Add(new DiscordMessageFile(file.Key, file.Value, null));
-            }
-
-
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the mention to the mentions to parse, etc. at the execution of the webhook.
-        /// </summary>
-        /// <param name="mention">Mention to add.</param>
-        public override DiscordWebhookBuilder AddMention(IMention mention)
-        {
-            this._mentions.Add(mention);
-            return this;
-        }
-
-        /// <summary>
-        /// Adds the mentions to the mentions to parse, etc. at the execution of the webhook.
-        /// </summary>
-        /// <param name="mentions">Mentions to add.</param>
-        public override DiscordWebhookBuilder AddMentions(IEnumerable<IMention> mentions)
-        {
-            this._mentions.AddRange(mentions);
             return this;
         }
 
@@ -332,24 +109,6 @@ namespace DSharpPlus.Entities
         /// <param name="messageId">The id of the message to modify.</param>
         /// <returns>The modified message</returns>
         public async Task<DiscordMessage> ModifyAsync(DiscordWebhook webhook, ulong messageId) => await webhook.EditMessageAsync(messageId, this).ConfigureAwait(false);
-
-        /// <summary>
-        /// Clears all message components on this builder.
-        /// </summary>
-        public override void ClearComponents()
-            => this._components.Clear();
-
-        /// <summary>
-        /// Allows for clearing the Webhook Builder so that it can be used again to send a new message.
-        /// </summary>
-        public override void Clear()
-        {
-            this.Content = "";
-            this._embeds.Clear();
-            this.IsTTS = false;
-            this._mentions.Clear();
-            this._files.Clear();
-        }
 
         /// <summary>
         /// Does the validation before we send a the Create/Modify request.
